review: Address split view feedback
- Update outline-icons to 4.5.0 and use SplitIcon for the open in split
  view actions.
- Move toLocationDescriptor to utils/history alongside the other
  location helpers.
- Track the focused pane in an observable box instead of manually
  maintained listeners; observer components read it directly and
  imperative callers are unaffected.
