+++ package.json
@@ -182,7 +182,7 @@
     "node-fetch": "2.7.0",
     "nodemailer": "^9.0.1",
     "octokit": "^5.0.5",
-    "outline-icons": "^4.3.0",
+    "outline-icons": "^4.5.0",
     "oy-vey": "^0.12.1",
     "pako": "^2.1.0",
     "passport": "^0.7.0",
